+++ exporters/trace/stackdriver/src/main/java/io/opencensus/exporter/trace/stackdriver/StackdriverTraceConfiguration.java
@@ -20,8 +20,10 @@ import com.google.auth.Credentials;
 import com.google.auto.value.AutoValue;
 import com.google.cloud.ServiceOptions;
 import com.google.cloud.trace.v2.stub.TraceServiceStub;
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
+import io.opencensus.common.Duration;
 import io.opencensus.trace.AttributeValue;
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -40,6 +42,8 @@ public abstract class StackdriverTraceConfiguration {
 
   private static final String DEFAULT_PROJECT_ID =
       Strings.nullToEmpty(ServiceOptions.getDefaultProjectId());
+
+  @VisibleForTesting static final Duration DEFAULT_DEADLINE = Duration.create(10, 0);
 
   StackdriverTraceConfiguration() {}
 
@@ -78,6 +82,16 @@ public abstract class StackdriverTraceConfiguration {
   public abstract Map<String, AttributeValue> getFixedAttributes();
 
   /**
+   * Returns the deadline for exporting to Stackdriver Trace backend.
+   *
+   * <p>Default value is 10 seconds.
+   *
+   * @return the export deadline.
+   * @since 0.22
+   */
+  public abstract Duration getDeadline();
+
+  /**
    * Returns a new {@link Builder}.
    *
    * @return a {@code Builder}.
@@ -86,7 +100,8 @@ public abstract class StackdriverTraceConfiguration {
   public static Builder builder() {
     return new AutoValue_StackdriverTraceConfiguration.Builder()
         .setProjectId(DEFAULT_PROJECT_ID)
-        .setFixedAttributes(Collections.<String, AttributeValue>emptyMap());
+        .setFixedAttributes(Collections.<String, AttributeValue>emptyMap())
+        .setDeadline(DEFAULT_DEADLINE);
   }
 
   /**
@@ -96,6 +111,8 @@ public abstract class StackdriverTraceConfiguration {
    */
   @AutoValue.Builder
   public abstract static class Builder {
+
+    @VisibleForTesting static final Duration ZERO = Duration.fromMillis(0);
 
     Builder() {}
 
@@ -135,9 +152,23 @@ public abstract class StackdriverTraceConfiguration {
      */
     public abstract Builder setFixedAttributes(Map<String, AttributeValue> fixedAttributes);
 
+    /**
+     * Sets the deadline for exporting to Stackdriver Trace backend.
+     *
+     * <p>If both {@code TraceServiceStub} and {@code Deadline} are set, {@code TraceServiceStub}
+     * takes precedence and {@code Deadline} will not be respected.
+     *
+     * @param deadline the export deadline.
+     * @return this
+     * @since 0.22
+     */
+    public abstract Builder setDeadline(Duration deadline);
+
     abstract String getProjectId();
 
     abstract Map<String, AttributeValue> getFixedAttributes();
+
+    abstract Duration getDeadline();
 
     abstract StackdriverTraceConfiguration autoBuild();
 
@@ -159,6 +190,7 @@ public abstract class StackdriverTraceConfiguration {
         Preconditions.checkNotNull(fixedAttribute.getKey(), "attribute key");
         Preconditions.checkNotNull(fixedAttribute.getValue(), "attribute value");
       }
+      Preconditions.checkArgument(getDeadline().compareTo(ZERO) > 0, "Deadline must be positive.");
       return autoBuild();
     }
   }